requirements.txt
cinemagoer==2022.12.4
imdb-sqlite==0.1.3
jupyterlab==3.5.1
matplotlib==3.6.2
pandas==1.5.2
